+++ src/flat.rs
@@ -6,7 +6,7 @@ use crate::parallel::*;
 
 /// Projects all geographic points onto a flat surface for faster geodesic calculation
 ///
-pub fn to_flat_points<T: Point>(points: &[T]) -> Vec<FlatPoint<f64>> {
+pub fn to_flat_points<T: Point>(points: &[T]) -> Vec<FlatPoint<f32>> {
     let center = points.center_lat().unwrap();
     let proj = FlatProjection::new(center);
 
@@ -16,11 +16,11 @@ pub fn to_flat_points<T: Point>(points: &[T]) -> Vec<FlatPoint<f64>> {
 }
 
 trait CenterLatitude {
-    fn center_lat(self: &Self) -> Option<f64>;
+    fn center_lat(self: &Self) -> Option<f32>;
 }
 
 impl<T: Point> CenterLatitude for [T] {
-    fn center_lat(self: &Self) -> Option<f64> {
+    fn center_lat(self: &Self) -> Option<f32> {
         let lat_min = self.iter().map(|fix| fix.latitude()).ord_subset_min()?;
         let lat_max = self.iter().map(|fix| fix.latitude()).ord_subset_max()?;
